Move impl to struct
